pyproject.toml
[project]
name = "y-o-p-o"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "ultralytics>=8.3.217",
    "uvicorn>=0.38.0",
]
